Added support for custom scrolling context
This change makes it possible to use scrollToBySpeed for elements with scrollable content (e.g., overflow: auto), instead of only on the whole window. It's working great in my particular use case (fixed position full screen divs that we switch between using off canvas layout techniques), but I haven't tested this outside of that, and I have not tested this with the default (html,body) usage. It might still need a little bit of tweaking, but I think this is a useful feature so I thought it was worth sharing the code.